Refactor in a package
