+++ pytest_fauxfactory/__init__.py
@@ -1,0 +1,1 @@
+# pytest_fauxfactory package
